Move utils inside shared folder
