+++ src/components/computer-mode/child-components/WordModal.vue
@@ -18,7 +18,7 @@
 </template>
 
 <script>
-import { randomNumber } from "@/utils";
+import { randomNumber } from "@/shared/utils";
 // Components
 import EnterWord from "./EnterWord";
 import HideChars from "./HideChars";
